Backend_-School_management_system/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
# backports.zoneinfo==0.2.1
certifi==2025.10.5
cffi>=2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
cssselect2==0.7.0
defusedxml==0.7.1
Django==5.2.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
html5lib==1.1
idna==3.11
lxml==6.0.2
oscrypto==1.3.0
pillow==10.4.0
pycparser==2.23
pyHanko==0.27.1
pyhanko-certvalidator==0.26.8
PyJWT==2.9.0
pypdf==5.9.0
pypng==0.20220715.0
python-bidi==0.6.7
PyYAML==6.0.3
qrcode==7.4.2
reportlab==4.4.3
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.13.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
webencodings==0.5.1
xhtml2pdf==0.2.17
django-cors-headers
openpyxl